Correción de bugs menores y cambios iniciales en el servidor para la generación de la evaluación basada en las exigencias del profesor.

+++ Hera/Services/FileServices/FileManagerService.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Hera.Data;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -9,10 +10,11 @@ namespace Hera.Services
     public class FileManagerService
     {
         private DirectoryInfo _directory;
+        private ApplicationDbContext _context;
 
-        public FileManagerService()
+        public FileManagerService(ApplicationDbContext context)
         {
-            
+            _context = context;
         }
 
         public void InitializePath()
@@ -25,9 +27,10 @@ namespace Hera.Services
             return Guid.NewGuid().ToString();
         }
 
-        public void DeleteFile(string filePath)
+        public void DeleteFile(string filePath, bool forced = false)
         {
-            if (File.Exists(filePath))
+            if (File.Exists(filePath)
+                && (Is_referenceFree(filePath) ) )
             {
                 File.Delete(filePath);
             }
@@ -43,6 +46,13 @@ namespace Hera.Services
                 DeleteFile(item);
             }
 
+        }
+
+        private bool Is_referenceFree(string filePath)
+        {
+            return _context.Desafios
+                .Where(d => d.DirSolucion.Equals(filePath))
+                .Count() <= 1;
         }
     }
 }
